requirements.txt
Pillow==9.5.0
Django==4.2
django-phonenumber-field==6.2.0
environs[django]==9.5.0
django-debug-toolbar==4.0.0
gunicorn==20.0.2
psycopg2-binary==2.9.5
django-widget-tweaks==1.4.12
django-phonenumbers==1.0.1
django-admin-charts==1.3.0
yookassa==2.3.5
urllib3==1.26.16